% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCarType.R
\name{carType.insert.callback}
\alias{carType.insert.callback}
\title{定义插入数据类型}
\usage{
carType.insert.callback(data, row, table = "t_md_carType",
  f = getCarType, id_var = "FCartypeName")
}
\arguments{
\item{data}{数据}

\item{row}{行}

\item{table}{表}

\item{f}{函数}

\item{id_var}{ID}
}
\value{
返回值
}
\description{
定义插入数据类型
}
\examples{
carType.insert.callback()
}
