% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCarType.R
\name{carType.update.callback}
\alias{carType.update.callback}
\title{更新数据类型}
\usage{
carType.update.callback(
  data,
  olddata,
  row,
  table = "t_md_carType",
  f = getCarType,
  edit.cols = c("FCartypeName", "FBusiObj_like", "FBusiOjb_same"),
  id_var = "FCartypeName"
)
}
\arguments{
\item{data}{数据}

\item{olddata}{原始数据}

\item{row}{行}

\item{table}{数据表}

\item{f}{函数}

\item{edit.cols}{编辑列}

\item{id_var}{代码}
}
\value{
返回
}
\description{
更新数据类型
}
\examples{
carType.update.callback()
}
