% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCarType.R
\name{readCartype}
\alias{readCartype}
\title{读取车型信息}
\usage{
readCartype(file_cartype, heading = c("品牌", "车型",
  "业务对象（近义词）", "业务对象（同义词）"))
}
\arguments{
\item{file_cartype}{车型文件}

\item{heading}{列名}
}
\value{
返回数据
}
\description{
读取车型信息
}
\examples{
readCartype()
}
