% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCarType.R
\name{selectCarType}
\alias{selectCarType}
\title{从数据库中读取数据}
\usage{
selectCarType(conn = conn_rds("nsic"))
}
\arguments{
\item{conn}{数据库连接}
}
\value{
返回值
}
\description{
从数据库中读取数据
}
\examples{
selectCarType()
}
